@@ -42,107 +42,109 @@
     "missing-translations": "node missing-translations",
     "test": "ng test --single-run",
     "test:unit": "ng test",
-    "test:coverage": "ng test --single-run --code-coverage",
-    "test:e2e": "ionic-app-scripts build && protractor e2e/protractor.conf.js",
-    "postinstall": "webdriver-manager update"
+    "test:coverage": "ng test --single-run --code-coverage --watch=false --progress=false --browsers=ChromeHeadlessNoSandbox",
+    "test:e2e": "npm run webdriver-update-ci && ionic-app-scripts build && protractor e2e/protractor.conf.js",
+    "webdriver-update-ci": "./node_modules/.bin/webdriver-manager update --standalone false --gecko false --versions.chrome 2.43.0"
   },
   "dependencies": {
-    "@angular/common": "4.4.3",
-    "@angular/compiler": "4.4.3",
-    "@angular/compiler-cli": "4.4.3",
-    "@angular/core": "4.4.3",
-    "@angular/forms": "4.4.3",
-    "@angular/http": "4.4.3",
-    "@angular/platform-browser": "4.4.3",
-    "@angular/platform-browser-dynamic": "4.4.3",
-    "@ionic-native/clipboard": "^4.5.3",
-    "@ionic-native/core": "^4.5.3",
-    "@ionic-native/in-app-browser": "^4.5.3",
-    "@ionic-native/keyboard": "^4.5.3",
-    "@ionic-native/network": "^4.5.3",
-    "@ionic-native/qr-scanner": "^4.5.3",
-    "@ionic-native/screen-orientation": "^4.5.3",
-    "@ionic-native/social-sharing": "^4.5.3",
-    "@ionic-native/splash-screen": "^4.5.3",
-    "@ionic-native/status-bar": "^4.5.3",
-    "@ionic-native/vibration": "^4.5.3",
-    "@ionic/storage": "^2.1.3",
+    "@angular/common": "^5.2.11",
+    "@angular/compiler": "^5.2.11",
+    "@angular/compiler-cli": "^5.2.11",
+    "@angular/core": "^5.2.11",
+    "@angular/forms": "^5.2.11",
+    "@angular/http": "^5.2.11",
+    "@angular/platform-browser": "^5.2.11",
+    "@angular/platform-browser-dynamic": "^5.2.11",
+    "@ionic-native/background-mode": "^4.17.0",
+    "@ionic-native/clipboard": "^4.16.0",
+    "@ionic-native/core": "^4.16.0",
+    "@ionic-native/in-app-browser": "^4.16.0",
+    "@ionic-native/keyboard": "^4.16.0",
+    "@ionic-native/local-notifications": "^4.17.0",
+    "@ionic-native/network": "^4.16.0",
+    "@ionic-native/qr-scanner": "^4.16.0",
+    "@ionic-native/screen-orientation": "^4.16.0",
+    "@ionic-native/social-sharing": "^4.16.0",
+    "@ionic-native/splash-screen": "^4.16.0",
+    "@ionic-native/status-bar": "^4.16.0",
+    "@ionic-native/vibration": "^4.16.0",
+    "@ionic/storage": "^2.2.0",
     "@ngx-translate/core": "^8.0.0",
     "@ngx-translate/http-loader": "2.0.0",
-    "@types/bcryptjs": "^2.4.1",
-    "@types/lodash": "^4.14.100",
-    "@types/moment-timezone": "^0.5.3",
-    "@types/uuid": "^3.4.3",
     "android-versions": "^1.3.0",
     "angular2-qrcode": "^2.0.1",
-    "ark-ts": "^0.3.7",
+    "ark-ts": "^0.3.16",
     "arkjs": "github:arkecosystem/ark-js#master",
     "bcryptjs": "^2.4.3",
     "bignumber.js": "^6.0.0",
     "bip38": "^2.0.2",
     "bip39": "^2.5.0",
-    "chart.js": "^2.7.1",
-    "cordova-android": "7.0.0",
-    "cordova-browser": "^5.0.3",
-    "cordova-clipboard": "^1.1.1",
-    "cordova-custom-config": "^5.0.2",
-    "cordova-ios": "4.5.4",
+    "chart.js": "^2.7.2",
+    "cordova-allow-background-audio": "^1.0.2",
+    "cordova-android": "6.3.0",
+    "cordova-browser": "5.0.4",
+    "cordova-clipboard": "^1.2.1",
+    "cordova-custom-config": "^5.1.0",
+    "cordova-ios": "4.5.5",
+    "cordova-plugin-add-swift-support": "^1.7.2",
     "cordova-plugin-compat": "^1.2.0",
     "cordova-plugin-console": "^1.1.0",
-    "cordova-plugin-device": "^2.0.1",
-    "cordova-plugin-inappbrowser": "^2.0.2",
-    "cordova-plugin-ionic-webview": "^1.1.16",
+    "cordova-plugin-device": "^2.0.2",
+    "cordova-plugin-inappbrowser": "^3.0.0",
+    "cordova-plugin-ionic-keyboard": "^2.1.3",
     "cordova-plugin-network-information": "^2.0.1",
-    "cordova-plugin-qrscanner": "^2.5.0",
-    "cordova-plugin-screen-orientation": "~3.0.1",
+    "cordova-plugin-qrscanner": "^2.6.2",
+    "cordova-plugin-screen-orientation": "^3.0.1",
     "cordova-plugin-splashscreen": "^5.0.2",
-    "cordova-plugin-statusbar": "^2.4.1",
+    "cordova-plugin-statusbar": "^2.4.2",
     "cordova-plugin-swift-support": "^3.1.1",
-    "cordova-plugin-vibration": "^3.0.1",
+    "cordova-plugin-vibration": "^3.1.0",
     "cordova-plugin-whitelist": "^1.3.3",
-    "cordova-plugin-x-socialsharing": "^5.2.1",
-    "cordova-sqlite-storage": "^2.2.0",
+    "cordova-plugin-x-socialsharing": "^5.4.4",
+    "cordova-sqlite-storage": "git+https://github.com/litehelpers/Cordova-sqlite-storage.git",
     "es6-promise-plugin": "^4.2.2",
     "iod": "^1.1.1",
     "ionic-angular": "^3.9.2",
-    "ionic-plugin-keyboard": "^2.2.1",
-    "ionic2-auto-complete": "^1.5.3-beta",
+    "ionic2-auto-complete": "^1.6.2-alpha",
     "ionicons": "3.0.0",
     "json-typescript-mapper": "github:luciorubeens/json-typescript-mapper#master",
-    "moment-timezone": "^0.5.14",
+    "moment-timezone": "^0.5.21",
     "ng2-charts": "^1.6.0",
-    "ngx-filter-pipe": "^1.0.1",
-    "node-forge": "^0.7.1",
-    "node-gyp": "^3.6.2",
-    "rxjs": "^5.5.6",
+    "ngx-filter-pipe": "^1.0.2",
+    "node-forge": "^0.7.6",
+    "rxjs": "^5.5.11",
+    "string-hash": "^1.1.3",
     "sw-toolbox": "3.6.0",
     "wif": "^2.0.6",
-    "zone.js": "0.8.20"
+    "zone.js": "0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "^1.6.7",
-    "@angular/router": "^5.2.3",
-    "@ionic/app-scripts": "3.1.8",
-    "@types/jasmine": "^2.8.6",
+    "@angular/cli": "^1.7.4",
+    "@angular/router": "^5.2.11",
+    "@ionic/app-scripts": "3.2.0",
+    "@types/bcryptjs": "^2.4.2",
+    "@types/jasmine": "^2.8.9",
+    "@types/lodash": "^4.14.117",
+    "@types/moment-timezone": "^0.5.9",
     "@types/node": "^8.9.4",
-    "angular-cli": "^1.0.0-beta.28.3",
-    "connect": "^3.6.5",
-    "ionic-mocks": "^1.2.1",
+    "@types/uuid": "^3.4.4",
+    "connect": "^3.6.6",
+    "ionic-mocks": "^1.3.0",
     "jasmine-core": "2.99.1",
-    "jasmine-reporters": "^2.3.0",
+    "jasmine-reporters": "^2.3.2",
     "jasmine-spec-reporter": "^4.2.1",
-    "karma": "^2.0.0",
+    "karma": "^3.0.0",
     "karma-chrome-launcher": "^2.2.0",
     "karma-cli": "^1.0.1",
-    "karma-coverage-istanbul-reporter": "^1.4.1",
+    "karma-coverage-istanbul-reporter": "^2.0.4",
     "karma-jasmine": "^1.1.1",
-    "karma-jasmine-html-reporter": "^0.2.2",
+    "karma-jasmine-html-reporter": "^1.3.1",
     "karma-spec-reporter": "0.0.32",
-    "karma-webpack": "^2.0.9",
-    "protractor": "^5.3.0",
+    "karma-webpack": "^3.0.5",
+    "protractor": "^5.4.1",
     "serve-static": "^1.13.1",
     "ts-node": "^4.1.0",
-    "typescript": "^2.6.2"
+    "typescript": "^2.7.2"
   },
   "description": "A Wallet for Everyone",
   "config": {
