@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { HttpClient, HttpHeaders } from '@angular/common/http';
+import { HttpClient } from '@angular/common/http';
 
 import { Observable } from 'rxjs/Observable';
 import { Subject } from 'rxjs/Subject';
@@ -9,15 +9,22 @@ import { UserDataProvider } from '@providers/user-data/user-data';
 import { StorageProvider } from '@providers/storage/storage';
 import { ToastProvider } from '@providers/toast/toast';
 
-const packageJson = require('@root/package.json');
-import { Transaction, TranslatableObject } from '@models/model';
+import { Transaction, TranslatableObject, BlocksEpochResponse } from '@models/model';
 
 import * as arkts from 'ark-ts';
 import lodash from 'lodash';
+import moment from 'moment';
 import * as constants from '@app/app.constants';
 import arktsConfig from 'ark-ts/config';
 import { ArkUtility } from '../../utils/ark-utility';
 import { Delegate } from 'ark-ts';
+import { StoredNetwork, FeeStatistic } from '@models/stored-network';
+
+interface NodeConfigurationResponse {
+  data: {
+    feeStatistics: FeeStatistic[]
+  };
+}
 
 @Injectable()
 export class ArkApiProvider {
@@ -26,35 +33,26 @@ export class ArkApiProvider {
   public onUpdateDelegates$: Subject<arkts.Delegate[]> = new Subject<arkts.Delegate[]>();
   public onSendTransaction$: Subject<arkts.Transaction> = new Subject<arkts.Transaction>();
 
-  private _network: arkts.Network;
+  private _network: StoredNetwork;
   private _api: arkts.Client;
 
   private _fees: arkts.Fees;
   private _delegates: arkts.Delegate[];
-  private _peers: arkts.Peer[];
 
   public arkjs = require('arkjs');
 
   constructor(
+    private httpClient: HttpClient,
     private userDataProvider: UserDataProvider,
     private storageProvider: StorageProvider,
-    private toastProvider: ToastProvider,
-    private http: HttpClient) {
+    private toastProvider: ToastProvider) {
     this.loadData();
 
     this.userDataProvider.onActivateNetwork$.subscribe((network) => {
+
       if (lodash.isEmpty(network)) { return; }
 
-      // set default peer
-      if (network && !network.activePeer) {
-        network.activePeer = arkts.Network.getDefault(network.type).activePeer;
-      }
-
-      this._network = network;
-      this.arkjs.crypto.setNetworkVersion(this._network.version);
-
-      this._api = new arkts.Client(this._network);
-      this.findGoodPeer();
+      this.setNetwork(network);
     });
   }
 
@@ -64,6 +62,12 @@ export class ArkApiProvider {
 
   public get api() {
     return this._api;
+  }
+
+  public get feeStatistics () {
+    if (!lodash.isUndefined(this._network.feeStatistics)) { return Observable.of(this._network.feeStatistics); }
+
+    return this.fetchFeeStatistics();
   }
 
   public get fees() {
@@ -78,29 +82,145 @@ export class ArkApiProvider {
     return this.fetchDelegates(constants.NUM_ACTIVE_DELEGATES * 2);
   }
 
-  public findGoodPeer(): void {
+  public setNetwork(network: StoredNetwork) {
+    // set default peer
+    if (network.type !== null) {
+      const activePeer = network.activePeer;
+      const apiNetwork = arkts.Network.getDefault(network.type);
+      if (apiNetwork) {
+        network = Object.assign<StoredNetwork, arkts.Network>(network, apiNetwork);
+      }
+      if (activePeer) {
+        network.activePeer = activePeer;
+      }
+    }
+
+    this._network = network;
+    this.arkjs.crypto.setNetworkVersion(this._network.version);
+
+    this._api = new arkts.Client(this._network);
+    this.findGoodPeer();
+
+    // Fallback if the fetchEpoch fail
+    this._network.epoch = arktsConfig.blockchain.date;
+    this.userDataProvider.onUpdateNetwork$.next(this._network);
+
+    this.fetchFees().subscribe();
+    this.fetchEpoch().subscribe();
+  }
+
+  public async findGoodPeer() {
     // Get list from active peer
-    this._api.peer.list().subscribe((response) => {
-      if (response) {
-        const port = +this._network.activePeer.port;
-        const filteredPeers = lodash.filter(response.peers, (peer) => {
-          return peer['status'] === 'OK' && peer['port'] === port && peer.ip !== this._network.activePeer.ip && peer.ip !== '127.0.0.1';
-        });
-        const sortHeight = lodash.orderBy(filteredPeers, ['height', 'delay'], ['desc', 'asc']);
-        this._peers = sortHeight;
-        this.updateNetwork(sortHeight[0]);
+    this._api.peer.list().subscribe(async (response) => {
+      if (response && await this.findGoodPeerFromList(response.peers)) {
+        return;
       } else {
-        this.toastProvider.error('API.PEER_LIST_ERROR');
+        await this.tryGetFallbackPeer();
       }
     },
-    // Get list from file
-    () => {
-      return arkts.PeerApi
-        .findGoodPeer(this._network)
-        .subscribe((peer) => this.updateNetwork(peer), () => {
-          this.toastProvider.error('API.PEER_LIST_ERROR');
-        });
+    async () => await this.tryGetFallbackPeer());
+  }
+
+  private async tryGetFallbackPeer() {
+    if (await this.findGoodPeerFromList(this.network.peerList)) {
+      return;
+    }
+
+    // Custom network
+    if (!this.network.type) {
+      return;
+    }
+
+    // try get a peer from the hardcoded ark-ts peerlist (only works for main and devnet)
+    arkts.PeerApi
+      .findGoodPeer(this._network)
+      .subscribe((peer) => this.updateNetwork(peer),
+        () => this.toastProvider.error('API.PEER_LIST_ERROR'));
+  }
+
+  private async findGoodPeerFromList(peerList: arkts.Peer[]) {
+    if (!peerList || !peerList.length) {
+      return false;
+    }
+    const port = +(this._network.p2pPort || this._network.activePeer.port);
+    // Force P2P port if specified in the network
+    if (this._network.p2pPort) {
+      for (const peer of peerList) {
+        peer.port = +this._network.p2pPort;
+      }
+    }
+    const preFilteredPeers = lodash.filter(peerList, (peer) => {
+      if (peer['status'] !== 'OK') {
+        return false;
+      }
+
+      if (peer['port'] !== port) {
+        return false;
+      }
+
+      if (peer.ip === this._network.activePeer.ip || peer.ip === '127.0.0.1') {
+        return false;
+      }
+
+      return true;
     });
+
+    let filteredPeers = [];
+    if (!this._network.isV2) {
+      filteredPeers = preFilteredPeers;
+    } else {
+      const configChecks = [];
+      for (const peer of preFilteredPeers) {
+        configChecks.push(this._api.peer.getVersion2Config(peer.ip, peer.port).toPromise());
+      }
+
+      const peerConfigResponses = await Promise.all(configChecks.map(p => p.catch(e => e)));
+      for (const peerId in peerConfigResponses) {
+        const config = peerConfigResponses[peerId];
+        if (config && config.data) {
+          const apiConfig: any = lodash.find(config.data.plugins, (_, key) => key.split('/').reverse()[0] === 'core-api');
+          if (apiConfig && apiConfig.enabled && apiConfig.port) {
+            const peer = preFilteredPeers[peerId];
+            peer.port = apiConfig.port;
+            filteredPeers.push(peer);
+          }
+        }
+      }
+    }
+
+    const missingHeights = [];
+    const missingHeightRequests = [];
+    for (const peerId in filteredPeers) {
+      const peer = filteredPeers[peerId];
+      if (!peer.height) {
+        missingHeights.push({
+          id: peerId,
+          peer
+        });
+        missingHeightRequests.push(this._api.loader.synchronisationStatus(`http://${peer.ip}:${peer.port}`).toPromise());
+      }
+    }
+
+    if (missingHeightRequests.length) {
+      const missingHeightResponses = await Promise.all(missingHeightRequests.map(p => p.catch(e => e)));
+      for (const peerId in missingHeightResponses) {
+        const response = missingHeightResponses[peerId];
+        if (response && response.height) {
+          const missingHeight = missingHeights[peerId];
+          const peer = missingHeight.peer;
+          peer.height = response.height;
+          filteredPeers[missingHeight.peerId] = peer;
+        }
+      }
+    }
+
+    const sortedPeers = lodash.orderBy(filteredPeers, ['height', 'delay'], ['desc', 'asc']);
+    if (!sortedPeers.length) {
+      return false;
+    }
+    this._network.peerList = sortedPeers;
+    this.updateNetwork(sortedPeers[0]);
+    return true;
   }
 
   public fetchDelegates(numberDelegatesToGet: number, getAllDelegates = false): Observable<arkts.Delegate[]> {
@@ -141,12 +261,15 @@ export class ArkApiProvider {
   public createTransaction(transaction: Transaction, key: string, secondKey: string, secondPassphrase: string): Observable<Transaction> {
     return Observable.create((observer) => {
       const configNetwork = arktsConfig.networks[this._network.name];
-      const jsNetwork = {
-        messagePrefix: configNetwork.name,
-        bip32: configNetwork.bip32,
-        pubKeyHash: configNetwork.version,
-        wif: configNetwork.wif,
-      };
+      let jsNetwork;
+      if (configNetwork) {
+        jsNetwork = {
+          messagePrefix: configNetwork.name,
+          bip32: configNetwork.bip32,
+          pubKeyHash: configNetwork.version,
+          wif: configNetwork.wif,
+        };
+      }
 
       if (!arkts.PublicKey.validateAddress(transaction.address, this._network)) {
         observer.error({
@@ -176,8 +299,11 @@ export class ArkApiProvider {
         return observer.complete();
       }
 
+      const epochTime = moment(this._network.epoch).utc().valueOf();
+      const now = moment().valueOf();
+      transaction.timestamp = Math.floor((now - epochTime) / 1000);
+
       transaction.signature = null;
-      transaction.signSignature = null;
       transaction.id = null;
 
       const keys = this.arkjs.crypto.getKeys(key, jsNetwork);
@@ -199,22 +325,21 @@ export class ArkApiProvider {
 
   public postTransaction(transaction: arkts.Transaction, peer: arkts.Peer = this._network.activePeer, broadcast: boolean = true) {
     return Observable.create((observer) => {
-      let headers = new HttpHeaders().set('Content-Type', 'application/json');
-      headers = headers.append('os', 'ark-mobile');
-      headers = headers.append('version', packageJson.version);
-      headers = headers.append('port', '1');
-      headers = headers.append('nethash', this._network.nethash);
-
-      const url = `http://${peer.ip}:${peer.port}/peer/transactions`;
-      const data = JSON.stringify({ transactions: [transaction] });
-      this.http.post(url, data, { headers }).subscribe((result: arkts.TransactionPostResponse) => {
-        if (result.success) {
+      const compressTransaction = JSON.parse(JSON.stringify(transaction));
+      this._api.transaction.post(compressTransaction, peer).subscribe((result: arkts.TransactionPostResponse) => {
+        if (this.isSuccessfulResponse(result)) {
           this.onSendTransaction$.next(transaction);
+
           if (broadcast) {
-            this.broadcastTransaction(transaction);
-            this.toastProvider.success('API.TRANSACTION_SENT');
+            if (!this._network.isV2) {
+              this.broadcastTransaction(transaction);
+            }
           }
+
           observer.next(transaction);
+          if (this._network.isV2 && !result.data.accept.length && result.data.broadcast.length) {
+            this.toastProvider.warn('TRANSACTIONS_PAGE.WARNING.BROADCAST');
+          }
         } else {
           if (broadcast) {
             this.toastProvider.error('API.TRANSACTION_FAILED');
@@ -223,7 +348,6 @@ export class ArkApiProvider {
         }
       }, (error) => observer.error(error));
     });
-
   }
 
   public getDelegateByPublicKey(publicKey: string): Observable<Delegate> {
@@ -237,11 +361,26 @@ export class ArkApiProvider {
                .map(response => response && response.success ? response.delegate : null);
   }
 
-  private broadcastTransaction(transaction: arkts.Transaction) {
-    const max = 10;
 
-    for (const peer of this._peers.slice(0, max)) {
-      this.postTransaction(transaction, peer, false).subscribe();
+  private isSuccessfulResponse (response) {
+    if (!this._network.isV2) {
+      return response.success && response.transactionIds;
+    } else {
+      const { data, errors } = response;
+      return data && data.invalid.length === 0 && errors === null;
+    }
+  }
+
+  private broadcastTransaction(transaction: arkts.Transaction) {
+    if (!this._network.peerList || !this._network.peerList.length) {
+      return;
+    }
+
+    for (const peer of this._network.peerList.slice(0, 10)) {
+      this.postTransaction(transaction, peer, false).subscribe(
+        null,
+        null
+      );
     }
   }
 
@@ -251,7 +390,7 @@ export class ArkApiProvider {
       this.onUpdatePeer$.next(peer);
     }
     // Save in localStorage
-    this.userDataProvider.updateNetwork(this.userDataProvider.currentProfile.networkId, this._network);
+    this.userDataProvider.addOrUpdateNetwork(this._network, this.userDataProvider.currentProfile.networkId);
     this._api = new arkts.Client(this._network);
 
     this.fetchDelegates(constants.NUM_ACTIVE_DELEGATES * 2).subscribe((data) => {
@@ -259,6 +398,29 @@ export class ArkApiProvider {
     });
 
     this.fetchFees().subscribe();
+    this.fetchFeeStatistics().subscribe();
+  }
+
+  private fetchFeeStatistics(): Observable<FeeStatistic[]> {
+    if (!this._network || !this._network.isV2) {
+      return Observable.empty();
+    }
+
+    return Observable.create((observer) => {
+      this.httpClient.get(`${this._network.getPeerAPIUrl()}/api/v2/node/configuration`).subscribe((response: NodeConfigurationResponse) => {
+        const data = response.data;
+        this._network.feeStatistics = data.feeStatistics;
+        observer.next(this._network.feeStatistics);
+      }, e => observer.error(e));
+    });
+  }
+
+  private fetchEpoch(): Observable<BlocksEpochResponse> {
+    return this.httpClient.get(`${this._network.getPeerAPIUrl()}/api/blocks/getEpoch`).map((response: BlocksEpochResponse) => {
+      this._network.epoch = new Date(response.epoch);
+      this.userDataProvider.onUpdateNetwork$.next(this._network);
+      return response;
+    });
   }
 
   private fetchFees(): Observable<arkts.Fees> {
