@@ -2,218 +2,79 @@
 // tslint:disable:object-literal-sort-keys
 Object.defineProperty(exports, "__esModule", { value: true });
 exports.default = {
-     networks: {
+  networks: {
     mainnet: {
-      /*bip32: {
-        private: 0x2bf4530,
-        public: 0x2bf4968,
-      },*/
+      bip32: {
+        private: 46090600,
+        public: 46089520,
+      },
       name: 'mainnet',
-      nethash: '313ea34c8eb705f79e7bc298b788417ff3f7116c9596f5c9875e769ee2f4ede1',
+      nethash: 'a28cfbd5475471d9c23186976b17a482138de2c6edfc7daf0919a159d2c524e6',
       token: 'KAPU',
       symbol: 'ʞ',
       version: 0x2D,
       explorer: 'http://explorer.kapu.one',
       wif: 0xaa,
+      p2pPort: 9701,
+      apiPort: 9702,
+      p2pVersion: '2.0.0',
+      isV2: true,
       activePeer: {
-        ip: '51.15.198.173',
-        port: 4600,
+        ip: '51.15.136.166',
+        port: 9702,
       },
       peers: [
-        '51.15.85.189:9700',
-        '198.199.101.234:9700',
-        '194.135.83.46:9700',
-        '195.181.243.59:9700',
-        '45.76.34.197:9700',
-        '66.70.138.38:9700',
-        '51.15.89.225:9700',
-        '194.135.83.88:9700',
-        '108.61.242.196:9700',
-        '37.187.180.168:9700',
-        '164.8.251.179:9700',
-        '51.15.210.177:9700',
-        '212.24.106.81:9700',
-        '45.32.159.192:9700',
-        '51.15.223.11:9700',
-        '94.176.236.51:4600',
-        '194.135.84.233:9700',
-        '51.15.221.100:4600',
-        '51.15.84.234:4600',
-        '80.209.239.253:9700',
-        '51.15.207.184:9700',
-        '45.76.46.24:9700',
-        '79.143.187.45:9700',
-        '51.15.92.176:9700',
-        '45.76.17.182:9700',
-        '51.15.46.213:9700',
-        '80.211.230.30:9700',
-        '45.63.60.94:9700',
-        '45.63.78.203:9700',
-        '104.236.138.83:9700',
-        '94.176.233.213:4600',
-        '163.172.131.25:9700',
-        '51.15.198.173:4600',
-        '51.15.201.243:4600',
-        '94.177.218.123:9700',
-        '45.76.39.82:9700',
-        '144.217.242.167:4600',
-        '94.176.233.210:4600',
-        '51.15.91.96:9700',
-        '51.15.84.139:4600',
-        '45.32.29.180:4600',
-        '108.61.119.143:9700',
-        '80.241.218.21:4600',
-        '51.15.201.56:4600',
-        '45.63.100.48:9700',
-        '45.32.180.7:9700',
-        '51.15.43.222:9700',
-        '45.76.83.111:9700',
-        '51.15.217.253:9700',
-        '94.176.238.173:4600',
-        '51.15.203.197:9700',
-        '51.15.215.113:4600',
-        '45.77.136.77:9700',
-        '51.15.93.241:9700',
-        '51.15.85.79:9700',
-        '212.24.111.186:9700',
-        '51.15.194.207:4600',
-        '212.237.60.84:9700',
-        '92.222.217.140:9700',
-        '162.243.25.65:9700',
-        '45.76.33.73:9700',
-        '145.239.90.228:4600',
-        '51.15.77.90:9700',
-        '51.15.75.77:9700',
-        '195.181.241.25:9700',
-        '138.197.100.48:9700',
-        '45.63.119.129:9700',
-        '144.217.92.15:9700',
-        '51.15.133.233:9700',
-        '51.15.90.215:9700',
-        '142.44.162.174:9700',
-        '45.77.139.167:9700',
+        {
+          ip: '51.15.136.166',
+          port: 9702
+        },
+        {
+          ip: '51.15.84.139',
+          port: 9702
+        },
+        {
+          ip: '13.80.106.203',
+          port: 9702
+        },
+        {
+          ip: '51.15.90.150',
+          port: 9702
+        },
+        {
+          ip: '51.15.113.226',
+          port: 9702
+        },
+        {
+          ip: '51.15.115.99',
+          port: 9702
+        },
+        {
+          ip: '137.74.20.83',
+          port: 9702
+        }
       ],
     },
     devnet: {
-      /*bip32: {
+      bip32: {
         public: 0x043587cf,
         private: 0x04358394,
-      },*/
+      },
       name: 'devnet',
       nethash: 'f1ef11be7a879671b3019a785c9a2c9dbd9d800b05644d26ad132275ffe2dd48',
       token: 'KAPU',
       symbol: 'Dʞ',
       version: 0x50,
       explorer: 'https://explorer.devnet.kapunode.net',
-      /*wif: 0xba,*/
+      wif: 0xba,
+      p2pPort: 9701,
+      apiPort: 9702,
+      p2pVersion: '2.0.0',
+      isV2: true,
       activePeer: {
         ip: '199.247.24.132',
         port: 4100,
       },
-      peers: [/*
-        '45.63.74.5:4002',
-        '167.114.29.50:4002',
-        '167.114.29.45:4002',
-        '167.114.43.35:4002',
-        '104.129.43.144:4002',
-        '167.114.29.58:4002',
-        '173.253.129.20:4002',
-        '62.113.246.106:4002',
-        '45.76.218.117:4002',
-        '158.69.193.185:4002',
-        '144.217.194.188:4002',
-        '167.114.29.57:4002',
-        '167.114.29.40:4002',
-        '167.114.29.63:4002',
-        '185.177.21.65:4002',
-        '167.114.29.62:4002',
-        '167.114.29.54:4002',
-        '45.76.139.131:4002',
-        '138.68.29.19:4002',
-        '216.174.63.138:4002',
-        '213.32.9.98:4002',
-        '137.74.168.191:4002',
-        '104.236.122.76:4002',
-        '167.114.29.47:4002',
-        '167.114.29.37:4002',
-        '45.63.42.140:4002',
-        '45.33.86.168:4002',
-        '108.61.209.209:4002',
-        '167.114.29.33:4002',
-        '107.170.196.22:4002',
-        '178.62.102.209:4002',
-        '145.239.88.3:4002',
-        '167.114.43.37:4002',
-        '145.239.88.2:4002',
-        '217.182.70.37:4002',
-        '185.5.54.0:4002',
-        '45.76.174.177:4002',
-        '216.174.63.137:4002',
-        '167.114.29.59:4002',
-        '104.238.165.129:4002',
-        '217.182.79.88:4002',
-        '167.114.29.61:4002',
-        '51.255.198.137:4002',
-        '167.114.29.38:4002',
-        '167.114.43.33:4002',
-        '45.77.117.174:4002',
-        '213.32.9.97:4002',
-        '193.70.1.222:4002',
-        '45.76.39.61:4002',
-        '204.10.184.45:4002',
-        '45.76.93.4:4002',
-        '167.114.29.46:4002',
-        '91.121.1.174:4002',
-        '165.227.59.129:4002',
-        '167.114.29.44:4002',
-        '139.99.156.231:4002',
-        '164.8.251.90:4002',
-        '167.114.29.56:4002',
-        '108.61.169.37:4002',
-        '104.238.189.78:4002',
-        '173.254.228.67:4002',
-        '204.10.184.46:4002',
-        '66.70.207.89:4002',
-        '172.110.20.39:4002',
-        '137.74.47.141:4002',
-        '167.114.29.41:4002',
-        '167.114.29.52:4002',
-        '167.114.43.41:4002',
-        '51.254.103.18:4002',
-        '167.114.43.47:4002',
-        '45.76.136.154:4002',
-        '167.114.29.39:4002',
-        '167.114.29.34:4002',
-        '138.197.165.165:4002',
-        '167.114.43.34:4002',
-        '145.239.88.101:4002',
-        '45.76.245.155:4002',
-        '167.114.43.43:4002',
-        '91.134.115.30:4002',
-        '46.231.204.207:4002',
-        '167.114.43.46:4002',
-        '138.197.29.215:4002',
-        '104.238.177.174:4002',
-        '109.154.47.137:4002',
-        '104.207.138.122:4002',
-        '138.68.21.196:4002',
-        '167.114.29.36:4002',
-        '165.227.150.45:4002',
-        '37.59.70.165:4002',
-        '94.177.218.138:4002',
-        '167.114.29.55:4002',
-        '51.254.201.227:4002',
-        '167.114.43.48:4002',
-        '104.154.82.122:4002',
-        '217.182.78.226:4002',
-        '207.154.220.137:4002',
-        '195.181.219.91:4002',
-        '192.155.91.248:4002',
-        '167.114.29.53:4002',
-        '43.224.34.53:4002',
-        '167.114.43.40:4002',
-        '138.197.206.43:4002',
-        */
+      peers: [
       ],
     },
   },
