@@ -1,4 +1,4 @@
-import {Component, OnDestroy, OnInit, ViewChild} from '@angular/core';
+import { Component, ElementRef, Renderer2, OnDestroy, OnInit, ViewChild } from '@angular/core';
 import { Platform, Config, Nav, MenuController, AlertController, App, Events } from 'ionic-angular';
 import { StatusBar } from '@ionic-native/status-bar';
 import { SplashScreen } from '@ionic-native/splash-screen';
@@ -11,6 +11,7 @@ import { UserDataProvider } from '@providers/user-data/user-data';
 import { SettingsDataProvider } from '@providers/settings-data/settings-data';
 import { ArkApiProvider } from '@providers/ark-api/ark-api';
 import { ToastProvider } from '@providers/toast/toast';
+// import { LocalNotificationsProvider } from '@providers/local-notifications/local-notifications';
 
 import { TranslateService } from '@ngx-translate/core';
 
@@ -48,6 +49,7 @@ export class MyApp implements OnInit, OnDestroy {
     private arkApiProvider: ArkApiProvider,
     private settingsDataProvider: SettingsDataProvider,
     private toastProvider: ToastProvider,
+    // private localNotificationsProvider: LocalNotificationsProvider,
     private menuCtrl: MenuController,
     private alertCtrl: AlertController,
     private config: Config,
@@ -55,12 +57,14 @@ export class MyApp implements OnInit, OnDestroy {
     private screenOrientation: ScreenOrientation,
     private app: App,
     private ionicNetwork: Network,
-    splashScreen: SplashScreen,
-    events: Events
+    private splashScreen: SplashScreen,
+    private events: Events,
+    public element: ElementRef,
+    private renderer: Renderer2
   ) {
 
     platform.ready().then(() => {
-      splashScreen.hide();
+      this.splashScreen.hide();
       menuCtrl.enable(false, 'sidebarMenu');
 
       this.initConfig();
@@ -75,17 +79,21 @@ export class MyApp implements OnInit, OnDestroy {
         this.openPage('LoginPage');
       });
 
-      events.subscribe('qrScanner:show', () => {
+      this.events.subscribe('qrScanner:show', () => {
         this.hideNav = true;
       });
-      events.subscribe('qrScanner:hide', () => {
+      this.events.subscribe('qrScanner:hide', () => {
         this.hideNav = false;
       });
 
-      this.settingsDataProvider.onUpdate$.subscribe(() => this.initTranslate());
+      this.settingsDataProvider.onUpdate$.subscribe(() => {
+        this.initTranslate();
+        this.initTheme();
+      });
     });
 
     this.initTranslate();
+    this.initTheme();
   }
 
   setBackButton() {
@@ -125,10 +133,10 @@ export class MyApp implements OnInit, OnDestroy {
     this.config.set('android', 'autoFocusAssist', 'delay');
 
     if (this.platform.is('cordova')) {
+      // this.localNotificationsProvider.init();
 
       if (this.platform.is('ios')) {
         this.statusBar.styleDefault();
-        this.keyboard.disableScroll(false);
       }
 
       if (this.platform.is('android')) {
@@ -156,7 +164,6 @@ export class MyApp implements OnInit, OnDestroy {
         }
       });
 
-      this.keyboard.hideKeyboardAccessoryBar(true);
       this.keyboard.onKeyboardShow().subscribe(() => document.body.classList.add('keyboard-is-open'));
       this.keyboard.onKeyboardHide().subscribe(() => document.body.classList.remove('keyboard-is-open'));
 
@@ -175,6 +182,16 @@ export class MyApp implements OnInit, OnDestroy {
         this.exitText = translations['EXIT_APP_TEXT'];
         this.signOutText = translations['SIGN_OUT_PROFILE_TEXT'];
       });
+    });
+  }
+
+  initTheme() {
+    this.settingsDataProvider.settings.subscribe(settings => {
+      if (settings.darkMode) {
+        this.renderer.addClass(this.element.nativeElement.parentNode, 'dark-theme');
+      } else {
+        this.renderer.removeClass(this.element.nativeElement.parentNode, 'dark-theme');
+      }
     });
   }
 
@@ -263,5 +280,4 @@ export class MyApp implements OnInit, OnDestroy {
     this.unsubscriber$.complete();
     this.authProvider.logout();
   }
-
 }
