@@ -9,6 +9,9 @@ export const WALLET_REFRESH_TRANSACTIONS_MILLISECONDS = 10 * 1000;
 export const WALLET_MIN_NUMBER_CONFIRMATIONS = 51;
 export const BIP39_DOCUMENTATION_URL = 'https://github.com/bitcoin/bips/blob/master/bip-0039.mediawiki';
 
+// TRANSACTION SEND
+export const TRANSACTION_ADDRESS_SIZE = 20;
+
 // STORAGE
 export const STORAGE_MARKET_TICKER = 'market_ticker';
 export const STORAGE_MARKET_HISTORY = 'market_history';
@@ -20,7 +23,7 @@ export const STORAGE_ACTIVE_PROFILE = 'active_profile';
 export const STORAGE_INTROSEEN = 'intro';
 export const STORAGE_DELEGATES = 'delegates';
 export const STORAGE_FEES = 'fees';
-export const STORAGE_MASTERPASSWORD = 'KAPUmasterpassword!';
+export const STORAGE_MASTERPASSWORD = 'masterpassword';
 export const STORAGE_AUTH_UNLOCK_TIMESTAMP = 'auth_timestamp';
 export const STORAGE_AUTH_ATTEMPTS = 'auth_attempts';
 
@@ -29,10 +32,7 @@ export const TOAST_HIDE_DELAY = 3000;
 export const TOAST_POSITION = 'bottom';
 
 // COIN MARKETCAP API
-//https://min-api.cryptocompare.com/data/histoday?fsym=KAPU&allData=true&tsym=
-export const API_MARKET_URL = "https://min-api.cryptocompare.com";
-export const API_MARKET_HISTORY_ENDPOINT = 'data/histoday?fsym=KAPU&allData=true&tsym=';
-export const API_MARKET_TICKER_ENDPOINT = 'data/pricemultifull?fsyms=KAPU&tsyms=';
+export const API_MARKET_URL = 'https://min-api.cryptocompare.com';
 
 // GENERIC
 export const VIBRATION_TIME_MS = 30;
@@ -45,5 +45,5 @@ export const PIN_ATTEMPTS_TIMEOUT_MILLISECONDS = 30 * 1000;
 
 // ARK
 export const PRIVACY_POLICY_URL = 'https://kapu.one/PrivacyPolicy.txt';
-export const URI_QRCODE_SCHEME_PREFIX = "kapu:";
+export const URI_QRCODE_SCHEME_PREFIX = 'kapu:';
 export const NUM_ACTIVE_DELEGATES = 51;
