@@ -5,6 +5,7 @@ import { Wallet, WalletKeys, Transaction, TranslatableObject } from '@models/mod
 import { TranslateService } from '@ngx-translate/core';
 
 import lodash from 'lodash';
+import { AddressCheckResult } from '@providers/address-checker/address-check-result';
 
 @Component({
   selector: 'confirm-transaction',
@@ -14,6 +15,7 @@ export class ConfirmTransactionComponent {
 
   @Input('wallet') wallet: Wallet;
 
+  @Output('onClosed') onClosed: EventEmitter<string> = new EventEmitter();
   @Output('onError') onError: EventEmitter<string> = new EventEmitter();
   @Output('onConfirm') onConfirm: EventEmitter<Transaction> = new EventEmitter();
 
@@ -24,19 +26,26 @@ export class ConfirmTransactionComponent {
     private translateService: TranslateService
   ) { }
 
-  open(transaction: any, keys: WalletKeys) {
-    transaction = new Transaction(this.wallet.address).deserialize(transaction);
+  open(transaction: any, keys: WalletKeys, addressCheckResult?: AddressCheckResult) {
+    transaction = new Transaction(this.wallet.address, this.arkApiProvider.network).deserialize(transaction);
 
     this.arkApiProvider.createTransaction(transaction, keys.key, keys.secondKey, keys.secondPassphrase)
       .subscribe((tx) => {
         const modal = this.modalCtrl.create('ConfirmTransactionModal', {
-          transaction: tx
-        }, { cssClass: 'inset-modal-send', enableBackdropDismiss: true });
+          transaction: tx,
+          addressCheckResult: addressCheckResult
+        }, { enableBackdropDismiss: true });
 
         modal.onDidDismiss((result) => {
-          if (lodash.isUndefined(result)) { return; }
+          if (lodash.isUndefined(result)) {
+            return this.onClosed.emit();
+          }
 
-          if (!result.status) { return this.presentWrongModal(result); }
+          if (!result.status) {
+            this.onClosed.emit();
+
+            return this.presentWrongModal(result);
+          }
 
           this.onConfirm.emit(tx);
 
@@ -54,7 +63,7 @@ export class ConfirmTransactionComponent {
 
         modal.present();
       }, (error: TranslatableObject) => {
-        this.translateService.get(error.key, error.parameters)
+        this.translateService.get(error.key || (error as any).message || error as any, error.parameters)
           .subscribe((errorMessage) => {
             this.onError.emit(errorMessage);
             this.presentWrongModal({
@@ -68,7 +77,7 @@ export class ConfirmTransactionComponent {
   presentWrongModal(response) {
     const responseModal = this.modalCtrl.create('TransactionResponsePage', {
       response
-    }, { cssClass: 'inset-modal-small' });
+    });
 
     responseModal.present();
   }
