@@ -1,12 +1,23 @@
 import { NgModule } from '@angular/core';
 import { KeyboardAttachDirective } from './keyboard-attach/keyboard-attach';
-import { MainnetOnlyDirective } from './mainnet-only/mainnet-only';
+import { MarketNetOnlyDirective } from './marketnet-only/marketnet-only';
 import { HeaderScrollerDirective } from './header-scroller/header-scroller';
 import { HideOnKeyboardOpenDirective } from './hide-on-keyboard-open/hide-on-keyboard-open';
 
 @NgModule({
-  declarations: [KeyboardAttachDirective, MainnetOnlyDirective, HeaderScrollerDirective, HideOnKeyboardOpenDirective],
+  declarations: [
+    KeyboardAttachDirective,
+    MarketNetOnlyDirective,
+    HeaderScrollerDirective,
+    HideOnKeyboardOpenDirective
+  ],
   imports: [],
-  exports: [KeyboardAttachDirective, MainnetOnlyDirective, HeaderScrollerDirective, HideOnKeyboardOpenDirective]
+  exports: [
+    KeyboardAttachDirective,
+    MarketNetOnlyDirective,
+    HeaderScrollerDirective,
+    HideOnKeyboardOpenDirective
+  ]
 })
-export class DirectivesModule {}
+export class DirectivesModule {
+}
