@@ -1,5 +1,6 @@
-import {Component, OnDestroy} from '@angular/core';
+import { Component, OnDestroy, NgZone } from '@angular/core';
 import { IonicPage, NavController, NavParams, ViewController, LoadingController } from 'ionic-angular';
+import { TranslateService } from '@ngx-translate/core';
 
 import { Subject } from 'rxjs/Subject';
 import 'rxjs/add/operator/takeUntil';
@@ -12,6 +13,9 @@ import { Transaction, MarketTicker, MarketCurrency } from '@models/model';
 import { Network } from 'ark-ts/model';
 
 import lodash from 'lodash';
+import { AddressCheckResult} from '@providers/address-checker/address-check-result';
+import { AddressCheckResultType } from '@providers/address-checker/address-check-result-type';
+import { ArkUtility } from '../../utils/ark-utility';
 
 @IonicPage()
 @Component({
@@ -23,9 +27,12 @@ export class ConfirmTransactionModal implements OnDestroy {
   public transaction: Transaction;
   public address: string;
 
+  public addressCheckResult: AddressCheckResult;
   public marketCurrency: MarketCurrency;
   public ticker: MarketTicker;
   public currentNetwork: Network;
+  public checkTypes = AddressCheckResultType;
+  public hasBroadcast = false;
 
   private unsubscriber$: Subject<void> = new Subject<void>();
 
@@ -37,8 +44,11 @@ export class ConfirmTransactionModal implements OnDestroy {
     private marketDataProvider: MarketDataProvider,
     private settingsDataProvider: SettingsDataProvider,
     private loadingCtrl: LoadingController,
+    private ngZone: NgZone,
+    private translateService: TranslateService,
   ) {
     this.transaction = this.navParams.get('transaction');
+    this.addressCheckResult = this.navParams.get('addressCheckResult');
     this.address = this.transaction.address;
 
     if (!this.transaction) { this.navCtrl.pop(); }
@@ -48,12 +58,38 @@ export class ConfirmTransactionModal implements OnDestroy {
   }
 
   broadcast() {
-    this.arkApiProvider.postTransaction(this.transaction)
-      .subscribe(() => {
+    if (this.hasBroadcast) {
+      return;
+    }
+
+    this.ngZone.run(() => {
+      this.hasBroadcast = true;
+      this.arkApiProvider.postTransaction(this.transaction).subscribe(() => {
         this.dismiss(true);
       }, (error) => {
-        this.dismiss(false, error.error);
+          this.translateService.get(
+            ['TRANSACTIONS_PAGE.ERROR.NOTHING_SENT', 'TRANSACTIONS_PAGE.ERROR.FEE_TOO_LOW'],
+            { fee: ArkUtility.subToUnit(this.transaction.fee) }
+          ).subscribe(translations => {
+            let message = error.message;
+
+            if (error.errors) {
+              const errors = error.errors || {};
+              const anyLowFee = Object.keys(errors).some(transactionId => {
+                return errors[transactionId].some(item => item.type === 'ERR_LOW_FEE');
+              });
+
+              if (anyLowFee) {
+                message = translations['TRANSACTIONS_PAGE.ERROR.FEE_TOO_LOW'];
+              } else {
+                message = translations['TRANSACTIONS_PAGE.ERROR.NOTHING_SENT'];
+              }
+            }
+
+            this.dismiss(false, message);
+          });
       });
+    });
   }
 
   dismiss(status?: boolean, message?: string) {
