@@ -3,14 +3,16 @@ import { IonicPageModule } from 'ionic-angular';
 
 import { TranslateModule } from '@ngx-translate/core';
 import {PassphraseWordTesterModal} from './passphrase-word-tester';
+import { PassphraseInputComponentModule } from '@components/passphrase-input/passphrase-input.module';
 
 @NgModule({
   declarations: [
-    PassphraseWordTesterModal,
+    PassphraseWordTesterModal
   ],
   imports: [
     IonicPageModule.forChild(PassphraseWordTesterModal),
-    TranslateModule
+    TranslateModule,
+    PassphraseInputComponentModule
   ],
 })
 export class PassphraseWordTesterModalModule {}
