@@ -1,5 +1,5 @@
-import {Component, OnDestroy} from '@angular/core';
-import { IonicPage, NavController, NavParams } from 'ionic-angular';
+import { Component, ViewChild, OnDestroy } from '@angular/core';
+import { AlertController, IonicPage, NavController, NavParams } from 'ionic-angular';
 
 import { Subject } from 'rxjs/Subject';
 import 'rxjs/add/operator/takeUntil';
@@ -7,8 +7,9 @@ import 'rxjs/add/operator/takeUntil';
 import { Profile } from '@models/profile';
 import { UserDataProvider } from '@providers/user-data/user-data';
 import { ToastProvider } from '@providers/toast/toast';
-
 import lodash from 'lodash';
+import { Network } from 'ark-ts/model';
+import { TranslateService } from '@ngx-translate/core';
 
 @IonicPage()
 @Component({
@@ -16,9 +17,13 @@ import lodash from 'lodash';
   templateUrl: 'profile-create.html',
 })
 export class ProfileCreatePage implements OnDestroy {
+  @ViewChild('createProfileForm') createProfileForm: HTMLFormElement;
 
-  public networks;
-  public networksIds;
+  public networks: {[networkId: string]: Network};
+  public networksIds: string[];
+  public networkChoices: {name: string, id?: string}[] = [];
+
+  public activeNetworkChoice: {name: string, id?: string};
 
   public newProfile = { name: '', networkId: '' };
   public showAdvancedOptions = false;
@@ -28,34 +33,60 @@ export class ProfileCreatePage implements OnDestroy {
   constructor(
     public navCtrl: NavController,
     public navParams: NavParams,
+    private alertCtrl: AlertController,
     private userDataProvider: UserDataProvider,
     private toastProvider: ToastProvider,
+    private translateService: TranslateService
   ) { }
 
-  onSelectNetwork(networkId: string) {
-    this.newProfile.networkId = networkId;
+  onSelectNetwork(networkChoice: {name: string, id?: string}) {
+    this.activeNetworkChoice = networkChoice;
+    this.newProfile.networkId = networkChoice.id;
   }
 
   submitForm() {
-    const profile = new Profile();
-    profile.name = this.newProfile.name;
-    profile.networkId = this.newProfile.networkId;
+    const existingProfile = this.userDataProvider.getProfileByName(this.newProfile.name);
 
-    this.userDataProvider.addProfile(profile).takeUntil(this.unsubscriber$).subscribe(() => {
-      this.navCtrl.setRoot('ProfileSigninPage');
-    }, () => {
-      this.toastProvider.error('PROFILES_PAGE.ADD_PROFILE_ERROR');
-    });
+    if (existingProfile) {
+      this.showAlert('PROFILES_PAGE.PROFILENAME_ALREADY_EXISTS', {name: this.newProfile.name});
+      this.createProfileForm.form.controls['name'].setErrors({incorrect: !!existingProfile});
+    } else {
+      const profile = new Profile();
+      profile.name = this.newProfile.name;
+      profile.networkId = this.newProfile.networkId;
+
+      this.userDataProvider.addProfile(profile).takeUntil(this.unsubscriber$).subscribe(() => {
+        this.navCtrl.setRoot('ProfileSigninPage');
+      }, () => {
+        this.toastProvider.error('PROFILES_PAGE.ADD_PROFILE_ERROR');
+      });
+    }
   }
 
   load() {
-    this.networks = this.userDataProvider.networks;
-    this.networksIds = lodash.keys(this.networks);
-    this.newProfile.networkId = this.networksIds[0];
+    this.translateService.get('PROFILES_PAGE.CUSTOM').subscribe(customTrans => {
+      this.networks = this.userDataProvider.networks;
+      this.networksIds = lodash.keys(this.networks);
+      this.networkChoices =
+        this.networksIds
+          .filter(id => this.userDataProvider
+                            .defaultNetworks
+                            .some(defaultNetwork => this.networks[id].nethash === defaultNetwork.nethash))
+          .map(id => {
+            return {name: this.networks[id].name, id: id};
+          });
+      this.networkChoices.push({name: customTrans, id: null});
+      this.newProfile.networkId = this.networksIds[0];
+      this.activeNetworkChoice = this.networkChoices[0];
+    });
   }
 
   toggleAdvanced() {
     this.showAdvancedOptions = !this.showAdvancedOptions;
+  }
+
+  public onCustomNetworkChange(customNetworkId: string) {
+    this.newProfile.networkId = customNetworkId;
   }
 
   ionViewDidLoad() {
@@ -67,4 +98,13 @@ export class ProfileCreatePage implements OnDestroy {
     this.unsubscriber$.complete();
   }
 
+  private showAlert(titleKey: string, stringParams: Object) {
+    this.translateService.get([titleKey, 'BACK_BUTTON_TEXT'], stringParams).subscribe((translation) => {
+      const alert = this.alertCtrl.create({
+        subTitle: translation[titleKey],
+        buttons: [translation.BACK_BUTTON_TEXT]
+      });
+      alert.present();
+    });
+  }
 }
