@@ -1,4 +1,4 @@
-import {Component, OnDestroy, ViewChild} from '@angular/core';
+import { Component, OnDestroy, ViewChild } from '@angular/core';
 import { IonicPage, NavController, NavParams, AlertController, ActionSheetController } from 'ionic-angular';
 
 import { Subject } from 'rxjs/Subject';
@@ -49,7 +49,7 @@ export class ProfileSigninPage implements OnDestroy {
       const buttons = [{
         text: translation.DELETE,
         role: 'delete',
-        icon: !this.platform.is('ios') ? 'ios-trash-outline' : '',
+        icon: this.platform.is('ios') ? 'ios-trash-outline' : 'md-trash',
         handler: () => {
           if (!this.profileHasWallets(profileId)) {
             this.showDeleteConfirm(profileId);
@@ -121,6 +121,9 @@ export class ProfileSigninPage implements OnDestroy {
 
     this.addresses = lodash(this.profiles).mapValues((o) => [o.name, o.networkId]).transform((result, data, id) => {
       const network = this.networks[data[1]];
+      if (!network) {
+        return;
+      }
       const networkName = lodash.capitalize(network.name);
       const isMainnet = network.type === NetworkType.Mainnet;
 
