@@ -1,40 +1,60 @@
-import { Component, OnInit, ViewChild } from '@angular/core';
-import { IonicPage, NavController, NavParams } from 'ionic-angular';
+import { Component, OnInit, ViewChild, NgZone } from '@angular/core';
+import { IonicPage, LoadingController, NavParams, Loading } from 'ionic-angular';
 import { FormGroup, Validators, FormControl } from '@angular/forms';
 
-import { Contact, Wallet, MarketTicker, MarketCurrency, SendTransactionForm, WalletKeys } from '@models/model';
+import { Contact, Wallet, SendTransactionForm, WalletKeys, QRCodeScheme } from '@models/model';
 
 import { UserDataProvider } from '@providers/user-data/user-data';
 import { ContactsProvider } from '@providers/contacts/contacts';
-import { MarketDataProvider } from '@providers/market-data/market-data';
-import { SettingsDataProvider } from '@providers/settings-data/settings-data';
 import { ArkApiProvider } from '@providers/ark-api/ark-api';
 import { ToastProvider } from '@providers/toast/toast';
 
-import { ContactsAutoCompleteService } from '@providers/contacts-auto-complete/contacts-auto-complete';
+import { AccountAutoCompleteService } from '@providers/account-auto-complete/account-auto-complete';
 
 import { PublicKey } from 'ark-ts/core';
 import { Network, Fees } from 'ark-ts/model';
+import { Subject } from 'rxjs/Subject';
 
-import { UnitsSatoshiPipe } from '@pipes/units-satoshi/units-satoshi';
+import { TruncateMiddlePipe } from '@pipes/truncate-middle/truncate-middle';
 import { PinCodeComponent } from '@components/pin-code/pin-code';
 import { ConfirmTransactionComponent } from '@components/confirm-transaction/confirm-transaction';
 import { QRScannerComponent } from '@components/qr-scanner/qr-scanner';
 import * as constants from '@app/app.constants';
-import { TransactionSend } from 'ark-ts';
+import { TransactionSend, TransactionType } from 'ark-ts';
 
 import { AutoCompleteComponent } from 'ionic2-auto-complete';
-import { AutoCompleteContact } from '@models/contact';
+import { AutoCompleteAccount, AutoCompleteAccountType } from '@models/contact';
 import { TranslatableObject } from '@models/translate';
-import { QRCodeScheme } from '@models/model';
 import { BigNumber } from 'bignumber.js';
 import { ArkUtility } from '../../../utils/ark-utility';
+import { AddressCheckerProvider} from '@providers/address-checker/address-checker';
+import { AddressCheckResult } from '@providers/address-checker/address-check-result';
+import { AmountComponent } from '@components/amount/amount';
+import { Amount } from '@components/amount/amount.model';
+import { TranslateService } from '@ngx-translate/core';
+
+class CombinedResult {
+  public checkerDone: boolean;
+  public checkerResult: AddressCheckResult;
+  public pinCodeDone: boolean;
+  public keys: WalletKeys;
+
+  public constructor(public loader: Loading) {
+  }
+}
+
+enum AddressType {
+  Unknown,
+  Contact,
+  WalletWithLabel,
+  WalletWithoutLabel
+}
 
 @IonicPage()
 @Component({
   selector: 'page-transaction-send',
   templateUrl: 'transaction-send.html',
-  providers: [UnitsSatoshiPipe],
+  providers: [TruncateMiddlePipe],
 })
 export class TransactionSendPage implements OnInit {
   @ViewChild('sendTransactionForm') sendTransactionHTMLForm: HTMLFormElement;
@@ -42,50 +62,61 @@ export class TransactionSendPage implements OnInit {
   @ViewChild('confirmTransaction') confirmTransaction: ConfirmTransactionComponent;
   @ViewChild('qrScanner') qrScanner: QRScannerComponent;
   @ViewChild('searchBar') searchBar: AutoCompleteComponent;
+  @ViewChild(AmountComponent) private amountComponent: AmountComponent;
 
   sendForm: FormGroup;
   transaction: SendTransactionForm = {};
 
   currentWallet: Wallet;
   currentNetwork: Network;
-  marketTicker: MarketTicker;
-  marketCurrency: MarketCurrency;
-  fees: Fees;
-  isExistingContact = true;
+  fee: number;
+  addressType: AddressType = AddressType.Unknown;
+  addressTypes = AddressType;
   isRecipientNameAutoSet: boolean;
-
-  tokenPlaceholder = 100;
-  fiatPlaceholder: number;
+  hasSent = false;
+  sendAllEnabled = false;
+  transactionType = TransactionType.SendArk;
 
   private currentAutoCompleteFieldValue: string;
+  private unsubscriber$: Subject<void> = new Subject<void>();
 
   constructor(
-    public navCtrl: NavController,
-    public navParams: NavParams,
+    private navParams: NavParams,
     private userDataProvider: UserDataProvider,
     private contactsProvider: ContactsProvider,
     private arkApiProvider: ArkApiProvider,
-    private marketDataProvider: MarketDataProvider,
-    private settingsDataProvider: SettingsDataProvider,
     private toastProvider: ToastProvider,
-    public contactsAutoCompleteService: ContactsAutoCompleteService,
-    private unitsSatoshiPipe: UnitsSatoshiPipe,
+    public contactsAutoCompleteService: AccountAutoCompleteService,
+    private truncateMiddlePipe: TruncateMiddlePipe,
+    private addressChecker: AddressCheckerProvider,
+    private loadingCtrl: LoadingController,
+    private translateService: TranslateService,
+    private ngZone: NgZone,
   ) {
     this.currentWallet = this.userDataProvider.currentWallet;
     this.currentNetwork = this.userDataProvider.currentNetwork;
   }
 
+  toggleSendAll () {
+    this.sendAllEnabled = !this.sendAllEnabled;
+
+    if (this.sendAllEnabled) {
+      this.sendAll();
+    }
+  }
+
   sendAll() {
     const balance = Number(this.currentWallet.balance);
-    const sendableAmount = balance - this.fees.send;
+    const sendableAmount = balance - this.fee;
+
     if (sendableAmount <= 0) {
       this.toastProvider.error({
           key: 'API.BALANCE_TOO_LOW_DETAIL',
           parameters: {
             token: this.currentNetwork.token,
-            fee: ArkUtility.arktoshiToArk(this.fees.send),
+            fee: ArkUtility.arktoshiToArk(this.fee),
             amount: ArkUtility.arktoshiToArk(balance),
-            totalAmount: ArkUtility.arktoshiToArk(balance + this.fees.send),
+            totalAmount: ArkUtility.arktoshiToArk(balance + this.fee),
             balance: ArkUtility.arktoshiToArk(balance)
           }
         } as TranslatableObject,
@@ -93,8 +124,8 @@ export class TransactionSendPage implements OnInit {
       return;
     }
 
-    this.transaction.amount = this.unitsSatoshiPipe.transform(sendableAmount, true);
-    this.onInputToken();
+    this.transaction.amount = ArkUtility.arktoshiToArk(sendableAmount, true);
+    this.amountComponent.setAmount(this.transaction.amount);
   }
 
   send() {
@@ -103,26 +134,46 @@ export class TransactionSendPage implements OnInit {
     } else if (!this.validAddress()) {
       this.toastProvider.error('TRANSACTIONS_PAGE.INVALID_ADDRESS_ERROR');
     } else {
-      this.createContact();
-      this.pinCode.open('PIN_CODE.TYPE_PIN_SIGN_TRANSACTION', true, true);
+      this.ngZone.run(() => {
+        this.hasSent = true;
+        this.createContactOrLabel();
+
+        this.translateService.get('TRANSACTIONS_PAGE.PERFORMING_DESTINATION_ADDRESS_CHECKS').subscribe(translation => {
+          const loader = this.loadingCtrl.create({content: translation});
+          const combinedResult: CombinedResult = new CombinedResult(loader);
+          this.addressChecker.checkAddress(this.transaction.recipientAddress).subscribe(checkerResult => {
+            combinedResult.checkerDone = true;
+            combinedResult.checkerResult = checkerResult;
+            this.createTransactionAndShowConfirm(combinedResult);
+          });
+          this.pinCode.open('PIN_CODE.TYPE_PIN_SIGN_TRANSACTION', true, true, (keys: WalletKeys) => {
+            combinedResult.pinCodeDone = true;
+            combinedResult.keys = keys;
+            this.createTransactionAndShowConfirm(combinedResult);
+          });
+        });
+      });
     }
   }
 
-  public onSearchItem(contact: AutoCompleteContact): void {
-    if (!contact || !contact.address) {
+  public hasNotSent(): void {
+    this.hasSent = false;
+  }
+
+  public onSearchItem(account: AutoCompleteAccount): void {
+    if (!account || !account.address) {
       return;
     }
 
-    this.transaction.recipientAddress = contact.address;
-    this.currentAutoCompleteFieldValue = contact.name;
+    this.transaction.recipientAddress = account.address;
+    this.currentAutoCompleteFieldValue = account.name;
     this.isRecipientNameAutoSet = true;
 
-    // if we have a real contact or a wallet with a label, we don't show the field to add a new contact
-    if (contact.name !== contact.address) {
-      this.isExistingContact = true;
-      this.transaction.recipientName = contact.name;
+    if (account.name !== account.address) {
+      this.addressType = account.type === AutoCompleteAccountType.Wallet ? AddressType.WalletWithLabel : AddressType.Contact;
+      this.transaction.recipientName = account.name;
     } else {
-      this.isExistingContact = false;
+      this.addressType = account.type === AutoCompleteAccountType.Wallet ? AddressType.WalletWithoutLabel : AddressType.Unknown;
       this.transaction.recipientName = null;
     }
   }
@@ -137,6 +188,17 @@ export class TransactionSendPage implements OnInit {
     this.setRecipientByAddress(input);
   }
 
+  public showFullAddress(): void {
+    // When field has focus, show full address
+    this.searchBar.setValue(this.transaction.recipientAddress);
+  }
+
+  public truncateAddressMiddle(): void {
+    // When field loses focus, use ellipses to show beginning and end of address
+    const addressString = this.transaction.recipientAddress;
+    this.searchBar.setValue(this.truncateMiddlePipe.transform(addressString, constants.TRANSACTION_ADDRESS_SIZE, addressString));
+  }
+
   private validAddress(): boolean {
     const isValid = PublicKey.validateAddress(this.transaction.recipientAddress, this.currentNetwork);
     this.sendTransactionHTMLForm.form.controls['recipientAddress'].setErrors({ incorrect: !isValid });
@@ -145,33 +207,35 @@ export class TransactionSendPage implements OnInit {
   }
 
   private validForm(): boolean {
-    let isValid: boolean;
-    for (const name in this.sendTransactionHTMLForm.form.controls) {
-      if (name === 'recipientAddress') {
-        continue;
-      }
-      const control = this.sendTransactionHTMLForm.form.controls[name];
-      isValid = control.valid;
-      if (!isValid) {
-        break;
-      }
+    let isValid = true;
+    if (
+      !this.sendTransactionHTMLForm.form.controls['amount'].value
+      || this.sendTransactionHTMLForm.form.controls['amount'].value <= 0
+      || this.sendTransactionHTMLForm.form.controls['smartBridge'].length > 64
+    ) {
+      isValid = false;
     }
 
     return isValid;
   }
 
-  createContact() {
-    if (this.isExistingContact || !this.transaction.recipientName) {
+  createContactOrLabel() {
+    if (this.addressType === AddressType.Contact || this.addressType === AddressType.WalletWithLabel || !this.transaction.recipientName) {
       return;
     }
 
     const validAddress = this.validAddress();
-    let validName = !!this.transaction.recipientName;
-    if (validName) {
-      validName = new RegExp('^[a-zA-Z0-9]+[a-zA-Z0-9- ]+$').test(this.transaction.recipientName);
-    }
+    const validName = new RegExp('^[a-zA-Z0-9]+[a-zA-Z0-9- ]+$').test(this.transaction.recipientName);
+
     if (validAddress && validName) {
-      this.contactsProvider.addContact(this.transaction.recipientName, this.transaction.recipientAddress);
+      if (this.addressType === AddressType.Unknown) {
+        this.contactsProvider.addContact(this.transaction.recipientAddress, this.transaction.recipientName).subscribe();
+      } else {
+        this.userDataProvider
+            .setWalletLabel(this.userDataProvider.getWalletByAddress(this.transaction.recipientAddress),
+                            this.transaction.recipientName)
+            .subscribe();
+      }
     }
   }
 
@@ -179,53 +243,67 @@ export class TransactionSendPage implements OnInit {
     this.qrScanner.open(true);
   }
 
-  onInputToken() {
-    const precision = this.marketCurrency.code === 'btc' ? 8 : 2;
-    this.transaction.amountEquivalent = +(this.transaction.amount * this.marketCurrency.price).toFixed(precision);
-  }
+  private createTransactionAndShowConfirm(result: CombinedResult) {
+    if (!result.pinCodeDone) {
+      return;
+    }
 
-  onInputFiat() {
-    this.transaction.amount = +(this.transaction.amountEquivalent / this.marketCurrency.price).toFixed(8);
-  }
+    if (!result.checkerDone) {
+      result.loader.present();
+      return;
+    }
 
-  onEnterPinCode(keys: WalletKeys) {
+    result.loader.dismiss();
+
     const amount = new BigNumber(this.transaction.amount);
     const data: TransactionSend = {
       amount: amount.times(constants.WALLET_UNIT_TO_SATOSHI).toNumber(),
       vendorField: this.transaction.smartBridge,
-      passphrase: keys.key,
-      secondPassphrase: keys.secondKey,
+      passphrase: result.keys.key,
+      secondPassphrase: result.keys.secondKey,
       recipientId: this.transaction.recipientAddress,
     };
 
     this.arkApiProvider.api.transaction.createTransaction(data).subscribe((transaction) => {
-      this.confirmTransaction.open(transaction, keys);
+      // The transaction will be signed again;
+      transaction.fee = this.fee;
+      this.confirmTransaction.open(transaction, result.keys, result.checkerResult);
     }, () => {
       this.toastProvider.error('TRANSACTIONS_PAGE.CREATE_TRANSACTION_ERROR');
+      this.hasNotSent();
     });
   }
 
   onScanQRCode(qrCode: QRCodeScheme) {
     if (qrCode.address) {
-      this.setFormValuesFromAddress(qrCode.address);
+      this.setFormValuesFromAddress(qrCode.address, qrCode.label);
+      const amount = Number(qrCode.amount);
+      if (!!amount) {
+        this.amountComponent.setAmount(amount);
+      }
+      if (qrCode.vendorField) {
+        this.transaction.smartBridge = qrCode.vendorField;
+      }
     } else {
       this.toastProvider.error('QR_CODE.INVALID_QR_ERROR');
     }
   }
 
   ionViewDidLoad() {
-    this.arkApiProvider.fees.subscribe((fees) => this.fees = fees);
-    this.marketDataProvider.ticker.subscribe((ticker) => {
-      this.marketTicker = ticker;
-
-      this.settingsDataProvider.settings.subscribe((settings) => {
-        this.marketCurrency = ticker.getCurrency({ code: settings.currency });
-        this.fiatPlaceholder = +(this.tokenPlaceholder * this.marketCurrency.price).toFixed(2);
-      });
-    });
+    this.hasNotSent();
   }
 
   ngOnInit(): void {
+    this.pinCode.onClosed.takeUntil(this.unsubscriber$).subscribe(() => {
+      this.hasNotSent();
+    });
+    this.confirmTransaction.onError.takeUntil(this.unsubscriber$).subscribe(() => {
+      this.hasNotSent();
+    });
+    this.confirmTransaction.onClosed.takeUntil(this.unsubscriber$).subscribe(() => {
+      this.hasNotSent();
+    });
+
     this.sendForm = new FormGroup({
       recipientAddress: new FormControl(''),
       recipientName: new FormControl(''),
@@ -239,30 +317,67 @@ export class TransactionSendPage implements OnInit {
     this.setFormValuesFromAddress(this.navParams.get('address') || '');
   }
 
-  private setFormValuesFromAddress(address: string): void {
+  ngOnDestroy() {
+    this.unsubscriber$.next();
+    this.unsubscriber$.complete();
+  }
+
+  public onAmountChange(newAmounts: Amount) {
+    this.transaction.amount = newAmounts.amount;
+    this.transaction.amountEquivalent = newAmounts.amountEquivalent;
+  }
+
+  public onFeeChange(newFee: number) {
+    this.fee = newFee;
+
+    if (this.sendAllEnabled) {
+      this.sendAll();
+    }
+  }
+
+  private setFormValuesFromAddress(address: string, alternativeRecipientName?: string): void {
     if (!address) {
       return;
     }
 
     this.sendForm.patchValue({recipientAddress: address});
-    this.setRecipientByAddress(address);
+    this.setRecipientByAddress(address, alternativeRecipientName);
+    this.truncateAddressMiddle();
   }
 
-  private setRecipientByAddress(input: string): void {
+  private setRecipientByAddress(input: string, alternativeRecipientName?: string): void {
+    if (input.indexOf('...') !== -1) {
+      return;
+    }
+
     this.currentAutoCompleteFieldValue = input;
     this.transaction.recipientAddress = input;
 
-    const contactOrLabel: Contact | string = this.contactsProvider.getContactByAddress(input)
-                                             || this.userDataProvider.getWalletLabel(input);
-    if (contactOrLabel && contactOrLabel !== input) {
-      this.isExistingContact = true;
+    const contact: Contact = this.contactsProvider.getContactByAddress(input);
+    if (contact) {
+      this.addressType = AddressType.Contact;
       this.isRecipientNameAutoSet = true;
-      this.transaction.recipientName = typeof contactOrLabel === 'string' ? contactOrLabel : contactOrLabel.name;
-    } else {
-      this.isExistingContact = false;
-      if (this.isRecipientNameAutoSet) {
-        this.transaction.recipientName = null;
-      }
+      this.transaction.recipientName = contact.name;
+      return;
+    }
+
+    const walletLabel = this.userDataProvider.getWalletLabel(input);
+    if (walletLabel) {
+      this.addressType = AddressType.WalletWithLabel;
+      this.isRecipientNameAutoSet = true;
+      this.transaction.recipientName = walletLabel;
+      return;
+    }
+
+    this.addressType = this.userDataProvider.getWalletByAddress(input)
+                         ? AddressType.WalletWithoutLabel
+                         : AddressType.Unknown;
+
+    if (alternativeRecipientName) {
+      this.isRecipientNameAutoSet = true;
+      this.transaction.recipientName = alternativeRecipientName;
+    } else if (this.isRecipientNameAutoSet) {
+      this.transaction.recipientName = null;
     }
   }
 }
