@@ -4,6 +4,7 @@ import { IonicPage, NavController, NavParams, ViewController } from 'ionic-angul
 import { Subject } from 'rxjs/Subject';
 import { ArkApiProvider } from '@providers/ark-api/ark-api';
 import lodash from 'lodash';
+import { TransactionType } from 'ark-ts/model';
 
 @IonicPage()
 @Component({
@@ -11,13 +12,13 @@ import lodash from 'lodash';
   templateUrl: 'register-delegate.html',
 })
 export class RegisterDelegatePage implements OnDestroy {
-
   public fee: number;
   public symbol: string;
   public name: string;
 
   public allowedDelegateNameChars = '[a-z0-9!@$&_.]+';
   public isExists = false;
+  public transactionType = TransactionType.CreateDelegate;
 
   private delegates;
   private unsubscriber$: Subject<void> = new Subject<void>();
@@ -28,7 +29,6 @@ export class RegisterDelegatePage implements OnDestroy {
     public viewCtrl: ViewController,
     private arkApiProvider: ArkApiProvider,
   ) {
-    this.arkApiProvider.fees.subscribe((fees) => this.fee = fees.delegate);
     this.symbol = this.arkApiProvider.network.symbol;
 
     this.arkApiProvider.delegates.takeUntil(this.unsubscriber$).subscribe((delegates) => this.delegates = delegates);
@@ -41,12 +41,16 @@ export class RegisterDelegatePage implements OnDestroy {
     this.isExists = !lodash.isNil(find);
   }
 
+  onFeeChange(newFee: number) {
+    this.fee = newFee;
+  }
+
   closeModal() {
     this.viewCtrl.dismiss();
   }
 
   submitForm() {
-    this.viewCtrl.dismiss(this.name);
+    this.viewCtrl.dismiss({ name: this.name, fee: this.fee });
   }
 
   ngOnDestroy() {
