@@ -22,14 +22,18 @@ export class SettingsDataProvider {
     languages: {
       'en': 'English',
       'de': 'Deutsch',
+      'fr': 'French',
+      'id': 'Indonesia',
       'it': 'Italiano',
       'nl': 'Nederlands',
       'sv': 'Svenska',
       'ru': 'Русский',
       'gr': 'Ελληνικά',
       'pt-br': 'Português',
+      'cs': 'Čeština',
       'kor': '한국어',
-      'bg': 'Български'
+      'bg': 'Български',
+      'pl': 'Polish'
     },
     currencies: {
       'btc': 'Bitcoin',
@@ -48,7 +52,18 @@ export class SettingsDataProvider {
       'hkd': 'Hong Kong Dollar',
       'idr': 'Indonesian Rupiah',
       'mxn': 'Mexican Peso',
+      'czk': 'Česká koruna'
     },
+    wordlistLanguages: {
+    'english': 'English',
+    'french': 'French',
+    'spanish': 'Spanish',
+    'italian': 'Italian',
+    'japanese': 'Japanese',
+    'korean': 'Korean',
+    'chinese_simplified': 'Chinese simplified',
+    'chinese_traditional': 'Chinese traditional'
+    }
   };
 
   constructor(private _storageProvider: StorageProvider, private translateService: TranslateService) {
